**Feature: Add favorite/un-favorite functionality to wallpaper preview**
**Description:**
- Implemented favorite and un-favorite actions directly in the wallpaper preview screen.
- Updated `ActionButtons` to toggle the favorite state dynamically.
- Integrated `FavouriteViewModel` and `ActionViewModel` for managing favorite actions.
- UI updates instantly to reflect the current favorite status.

**Benefits:**
- Improved user experience by enabling quick favorite/un-favorite actions within the preview.

+++ app/src/main/java/com/shahid/iqbal/screeny/data/local/dao/FavouriteWallpaperDao.kt
@@ -18,7 +18,7 @@ interface FavouriteWallpaperDao {
     suspend fun removeFromFavourite(wallpaper: FavouriteWallpaper)
 
     @Query("SELECT * FROM favourite_wallpaper WHERE id=:id limit 1")
-    suspend fun getFavouriteById(id:Int):FavouriteWallpaper?
+    suspend fun getFavouriteById(id:Long):FavouriteWallpaper?
 
 
     @Query("DELETE FROM favourite_wallpaper WHERE wallpaper=:url")
